+++ src/main/java/ru/khasang/cachoeira/controller/IController.java
@@ -17,7 +17,7 @@ public interface IController {
 
     void notifyChangeProject(IProject project);
     //добавил:
-    void handleAddTask(String nameOfTask, Date startDate, Date finishDate, Integer cost, double donePercent, PriorityList priorityType, List<IResource> resources);
+    void handleAddTask(String nameOfTask, Date startDate, Date finishDate, Double cost, double donePercent, PriorityList priorityType, List<IResource> resources);
 
     void notifyAddTask(ITask task);
 
@@ -25,7 +25,7 @@ public interface IController {
 
     void notifyRemoveTask(ITask task);
     //добавил:
-    void handleChangeTask(String taskNameField, Date taskStartDate, Date taskFinishDate, Integer taskCost, double taskDonePercent, PriorityList taskPriorityType, List<IResource> resources);
+    void handleChangeTask(String taskNameField, Date taskStartDate, Date taskFinishDate, Double taskCost, double taskDonePercent, PriorityList taskPriorityType, List<IResource> resources);
 
     void notifyChangeTask(ITask task);
